requirements.txt
nltk
newspaper3k
scikit-learn
lxml[html_clean]
flask
